Update README.md
Changed some wording...a question for you...
Is the name of the IP Block still IP Block for Nodes and Pods?  You reference and bold an IP Block for "K8s Cluster Networks".  It may be preferable to label it the same as you would see in the PKS Tile in OpsMan, but I was not sure if this was a Terraform nomenclature.

Also, I would change the image to remove the red-misspelling squiggly lines under Opsman etc.

There is also not mention of the SNAT rules to allow for BOSH to talk to vCenter and ESXi.  Should that be mentioned in the NAT configuration section?